fix(onboard): handle already-accepted/invalid invitation links
A Schulbegleiter who has already accepted their invite and reopens the
invitation link (the only URL they know) previously hit a bare
"Einladung ungültig" screen with no way forward.

- Gate /onboard on the session: any logged-in visitor is redirected to
  their start page (root route dispatches by role), so the invitee who
  reopens their link lands on their dashboard directly.
- For logged-out visitors, distinguish "bereits angenommen" from
  "ungültig oder abgelaufen" and add a "Zurück zur Startseite" button
  (→ /, which sends logged-out users to /login).
- Convert the page to a server component; extract the sign-up UI into a
  client OnboardForm and add an InvalidInvitationCard.
- Add InvitationFacade.getTokenState (non-throwing, distinguishes
  used/expired/notfound) and a getOnboardingState use-case; remove the
  now-unused fetchInviteDetails action + use-case.
